refactor panel contents into RouteInput
